Requests data columns by range: Move from initial sync to sync package.
Since it will eventually be used by the backfill package, and
the backfill packages does not depend on the initial sync package.
