+++ beacon-chain/sync/BUILD.bazel
@@ -157,7 +157,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    size = "small",
+    size = "medium",
     srcs = [
         "batch_verifier_test.go",
         "blobs_test.go",
