+++ beacon-chain/sync/data_columns.go
@@ -3,6 +3,7 @@ package sync
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/ethereum/go-ethereum/p2p/enode"
 	"github.com/libp2p/go-libp2p/core"
@@ -15,16 +16,29 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/startup"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/sync/verify"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/verification"
+	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
+	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
+	leakybucket "github.com/prysmaticlabs/prysm/v5/container/leaky-bucket"
 	eth "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/runtime/version"
+	"github.com/prysmaticlabs/prysm/v5/time/slots"
 	"github.com/sirupsen/logrus"
 )
 
-// RequestDataColumnSidecarsByRoot carefully selects, among `peers`,
-// the peers that can provide the requested data columns, requests them
-// and verify them according to `newColumnsVerifier` rules.
+// RequestDataColumnSidecarsByRoot is an opinionated, high level function which, for each data column in `dataColumnsToFetch`:
+//   - Greedily selects, among `peers`, the peers that can provide the requested data columns, to minimize the number of requests.
+//   - Request the data column sidecars from the selected peers.
+//   - In case of peers unable to actually provide all the requested data columns, retry with other peers.
+//
+// This function:
+//   - returns on success when all the initially missing sidecars in `dataColumnsToFetch` are retrieved, or
+//   - returns an error if all peers in `peers` are exhausted and at least one data column sidecar is still missing.
+//
+// TODO: In case at least one column is still missing after peer exhaustion,
+//
+//	but `peers` custody more than 64 columns, then try to fetch enough columns to reconstruct needed ones.
 func RequestDataColumnSidecarsByRoot(
 	ctx context.Context,
 	dataColumnsToFetch []uint64,
@@ -161,6 +175,189 @@ func RequestDataColumnSidecarsByRoot(
 
 	// If we still have remaining columns after all retries, return error
 	return nil, errors.Errorf("failed to retrieve all requested data columns after retries for block root=%#x, missing columns=%v", blockRoot, uint64MapToSortedSlice(remainingMissingColumns))
+}
+
+// RequestMissingDataColumnsByRange is an opinionated, high level function which, for each block in `blks`:
+//   - Computes all data column sidecars we should store and which are missing (according to our node ID and `groupCount`),
+//   - Builds an optimized set of data column sidecars by range requests in order to never request a data column that is already stored in the DB,
+//     and in order to minimize the number of total requests, while not exceeding `batchSize` sidecars per requests.
+//   - Greedily selects, among `peers`, the peers that can provide the requested data columns, to minimize the number of requests.
+//   - Request the data column sidecars from the selected peers.
+//   - In case of peers unable to actually provide all the requested data columns, retry with other peers.
+//
+// This function:
+//   - returns on success when all the initially missing sidecars for `blks` are retrieved, or
+//   - returns an error if no progress at all is made after 5 consecutives trials.
+//     (If at least one additional data column sidecar is retrieved between two trials, the counter is reset.)
+//
+// In case of success, initially missing data columns grouped by block root are returned.
+// This function expects blocks to be sorted by slot.
+//
+// TODO: In case at least one column is still missing after all allowed retries,
+//
+//	but `peers` custody more than 64 columns, then try to fetch enough columns to reconstruct needed ones.
+func RequestMissingDataColumnsByRange(
+	ctx context.Context,
+	clock *startup.Clock,
+	ctxMap ContextByteVersions,
+	p2p p2p.P2P,
+	rateLimiter *leakybucket.Collector,
+	groupCount uint64,
+	dataColumnsStorage filesystem.DataColumnStorageSummarizer,
+	peers []peer.ID,
+	blks []blocks.ROBlock,
+	batchSize int,
+) (map[[fieldparams.RootLength]byte][]blocks.RODataColumn, error) {
+	const maxAllowedStall = 5 // Number of trials before giving up.
+
+	if len(blks) == 0 {
+		return nil, nil
+	}
+
+	// Get the current slot.
+	currentSlot := clock.CurrentSlot()
+
+	// Compute the minimum slot for which we should serve data columns.
+	minimumSlot, err := DataColumnsRPCMinValidSlot(currentSlot)
+	if err != nil {
+		return nil, errors.Wrap(err, "data columns RPC min valid slot")
+	}
+
+	// Get blocks by root and compute all missing columns by root.
+	blockByRoot := make(map[[fieldparams.RootLength]byte]blocks.ROBlock, len(blks))
+	missingColumnsByRoot := make(map[[fieldparams.RootLength]byte]map[uint64]bool, len(blks))
+	for _, blk := range blks {
+		// Extract the block root and the block slot
+		blockRoot, blockSlot := blk.Root(), blk.Block().Slot()
+
+		// Populate the block by root.
+		blockByRoot[blockRoot] = blk
+
+		// Skip blocks that are not in the retention period.
+		if blockSlot < minimumSlot {
+			continue
+		}
+
+		missingColumns, err := MissingDataColumns(blk, p2p.NodeID(), groupCount, dataColumnsStorage)
+		if err != nil {
+			return nil, errors.Wrap(err, "missing data columns")
+		}
+
+		for _, column := range missingColumns {
+			if _, ok := missingColumnsByRoot[blockRoot]; !ok {
+				missingColumnsByRoot[blockRoot] = make(map[uint64]bool)
+			}
+			missingColumnsByRoot[blockRoot][column] = true
+		}
+	}
+
+	// Return early if there are no missing data columns.
+	if len(missingColumnsByRoot) == 0 {
+		return nil, nil
+	}
+
+	// Compute the number of missing data columns.
+	previousMissingDataColumnsCount := itemsCount(missingColumnsByRoot)
+
+	// Count the number of retries for the same amount of missing data columns.
+	stallCount := 0
+
+	// Add log fields.
+	log := log.WithFields(logrus.Fields{
+		"initialMissingColumnsCount": previousMissingDataColumnsCount,
+		"blockCount":                 len(blks),
+		"firstSlot":                  blks[0].Block().Slot(),
+		"lastSlot":                   blks[len(blks)-1].Block().Slot(),
+	})
+
+	// Log the start of the process.
+	start := time.Now()
+	log.Debug("Requesting data column sidecars - start")
+
+	alignedDataColumnsByRoot := make(map[[fieldparams.RootLength]byte][]blocks.RODataColumn, len(blks))
+	for len(missingColumnsByRoot) > 0 {
+		// Build requests.
+		requests, err := buildDataColumnByRangeRequests(blks, missingColumnsByRoot, batchSize)
+		if err != nil {
+			return nil, errors.Wrap(err, "build data column by range requests")
+		}
+
+		// Requests data column sidecars from peers.
+		retrievedDataColumnsByRoot := make(map[[fieldparams.RootLength]byte][]blocks.RODataColumn)
+		for _, request := range requests {
+			roDataColumns, err := fetchDataColumnsFromPeers(ctx, clock, p2p, rateLimiter, ctxMap, peers, request)
+			if err != nil {
+				return nil, errors.Wrap(err, "fetch data columns from peers")
+			}
+
+			for _, roDataColumn := range roDataColumns {
+				root := roDataColumn.BlockRoot()
+				if _, ok := blockByRoot[root]; !ok {
+					// It may happen if the peer which sent the data columns is on a different fork.
+					continue
+				}
+
+				retrievedDataColumnsByRoot[root] = append(retrievedDataColumnsByRoot[root], roDataColumn)
+			}
+		}
+
+		for root, dataColumns := range retrievedDataColumnsByRoot {
+			// Retrieve the block from the root.
+			block, ok := blockByRoot[root]
+			if !ok {
+				return nil, errors.New("block not found - this should never happen")
+			}
+
+			// Check if the data columns align with blocks.
+			if err := verify.DataColumnsAlignWithBlock(block, dataColumns); err != nil {
+				log.WithField("root", root).WithError(err).Debug("Data columns do not align with block")
+				continue
+			}
+
+			alignedDataColumnsByRoot[root] = append(alignedDataColumnsByRoot[root], dataColumns...)
+
+			// Remove aligned data columns from the missing columns.
+			for _, dataColumn := range dataColumns {
+				delete(missingColumnsByRoot[root], dataColumn.Index)
+				if len(missingColumnsByRoot[root]) == 0 {
+					delete(missingColumnsByRoot, root)
+				}
+			}
+		}
+
+		missingDataColumnsCount := itemsCount(missingColumnsByRoot)
+		if missingDataColumnsCount == previousMissingDataColumnsCount {
+			stallCount++
+		} else {
+			stallCount = 0
+		}
+
+		previousMissingDataColumnsCount = missingDataColumnsCount
+
+		if missingDataColumnsCount > 0 {
+			log := log.WithFields(logrus.Fields{
+				"remainingMissingColumnsCount": missingDataColumnsCount,
+				"stallCount":                   stallCount,
+				"maxAllowedStall":              maxAllowedStall,
+			})
+
+			if stallCount >= maxAllowedStall {
+				// It is very likely `bwbs` contains orphaned blocks, for which no peer has the data columns.
+				// We give up and let the state machine handle the situation.
+				const message = "Requesting data column sidecars - no progress, giving up"
+				log.Warning(message)
+				return nil, errors.New(message)
+			}
+
+			log.WithFields(logrus.Fields{
+				"remainingMissingColumnsCount": missingDataColumnsCount,
+				"stallCount":                   stallCount,
+			}).Debug("Requesting data column sidecars - continue")
+		}
+	}
+
+	log.WithField("duration", time.Since(start)).Debug("Requesting data column sidecars - success")
+	return alignedDataColumnsByRoot, nil
 }
 
 // MissingDataColumns looks at the data columns we should store for a given block regarding `custodyGroupCount`,
@@ -419,4 +616,352 @@ outerLoop:
 	}
 
 	return outputDataColumnsByPeer, descriptions
+}
+
+// buildDataColumnByRangeRequests builds an optimized slices of data column by range requests:
+// 1. It will never request a data column that is already stored in the DB if there is no "hole" in `roBlocks` other than missed slots.
+// 2. It will minimize the number of requests.
+// It expects blocks to be sorted by slot.
+func buildDataColumnByRangeRequests(roBlocks []blocks.ROBlock, missingColumnsByRoot map[[fieldparams.RootLength]byte]map[uint64]bool, batchSize int) ([]*eth.DataColumnSidecarsByRangeRequest, error) {
+	batchSizeSlot := primitives.Slot(batchSize)
+
+	// Return early if there are no blocks to process.
+	if len(roBlocks) == 0 {
+		return nil, nil
+	}
+
+	// It's safe to get the first item of the slice since we've already checked that it's not empty.
+	firstROBlock, lastROBlock := roBlocks[0], roBlocks[len(roBlocks)-1]
+	firstBlockSlot, lastBlockSlot := firstROBlock.Block().Slot(), lastROBlock.Block().Slot()
+	firstBlockRoot := firstROBlock.Root()
+
+	previousMissingDataColumns := make(map[uint64]bool, len(missingColumnsByRoot[firstBlockRoot]))
+
+	if missing, ok := missingColumnsByRoot[firstBlockRoot]; ok {
+		for key, value := range missing {
+			previousMissingDataColumns[key] = value
+		}
+	}
+
+	previousBlockSlot, previousStartBlockSlot := firstBlockSlot, firstBlockSlot
+
+	result := make([]*eth.DataColumnSidecarsByRangeRequest, 0, 1)
+	for index := 1; index < len(roBlocks); index++ {
+		roBlock := roBlocks[index]
+
+		// Extract the block from the RO-block.
+		block := roBlock.Block()
+
+		// Extract the slot from the block.
+		blockRoot, blockSlot := roBlock.Root(), block.Slot()
+
+		if blockSlot <= previousBlockSlot {
+			return nil, errors.Errorf("blocks are not strictly sorted by slot. Previous block slot: %d, current block slot: %d", previousBlockSlot, blockSlot)
+		}
+
+		// Extract KZG commitments count from the current block body
+		blockKzgCommitments, err := block.Body().BlobKzgCommitments()
+		if err != nil {
+			return nil, errors.Wrap(err, "blob KZG commitments")
+		}
+
+		// Compute the count of KZG commitments.
+		blockKzgCommitmentCount := len(blockKzgCommitments)
+
+		// Skip blocks without commitments.
+		if blockKzgCommitmentCount == 0 {
+			previousBlockSlot = blockSlot
+			continue
+		}
+
+		// Get the missing data columns for the current block.
+		missingDataColumns := make(map[uint64]bool, len(missingColumnsByRoot[blockRoot]))
+		for key, value := range missingColumnsByRoot[blockRoot] {
+			missingDataColumns[key] = value
+		}
+
+		// Compute if the missing data columns differ.
+		missingDataColumnsDiffer := uint64MapDiffer(previousMissingDataColumns, missingDataColumns)
+
+		// Compute if the batch size is reached.
+		batchSizeReached := blockSlot-previousStartBlockSlot >= batchSizeSlot
+
+		if missingDataColumnsDiffer || batchSizeReached {
+			// Append the slice to the result.
+			request := &eth.DataColumnSidecarsByRangeRequest{
+				StartSlot: previousStartBlockSlot,
+				Count:     uint64(blockSlot - previousStartBlockSlot),
+				Columns:   sortedSliceFromMap(previousMissingDataColumns),
+			}
+
+			result = append(result, request)
+
+			previousStartBlockSlot, previousMissingDataColumns = blockSlot, missingDataColumns
+		}
+
+		previousBlockSlot = blockSlot
+	}
+
+	lastRequest := &eth.DataColumnSidecarsByRangeRequest{
+		StartSlot: previousStartBlockSlot,
+		Count:     uint64(lastBlockSlot - previousStartBlockSlot + 1),
+		Columns:   sortedSliceFromMap(previousMissingDataColumns),
+	}
+
+	result = append(result, lastRequest)
+
+	return result, nil
+}
+
+// fetchDataColumnsFromPeers requests data columns by range to relevant peers
+func fetchDataColumnsFromPeers(
+	ctx context.Context,
+	clock *startup.Clock,
+	p2p p2p.P2P,
+	rateLimiter *leakybucket.Collector,
+	ctxMap ContextByteVersions,
+	peers []peer.ID,
+	targetRequest *eth.DataColumnSidecarsByRangeRequest,
+) ([]blocks.RODataColumn, error) {
+	// Filter out requests with no data columns.
+	if len(targetRequest.Columns) == 0 {
+		return nil, nil
+	}
+
+	// Get all admissible peers with the data columns they custody.
+	dataColumnsByAdmissiblePeer, err := waitForPeersForDataColumns(p2p, rateLimiter, peers, targetRequest)
+	if err != nil {
+		return nil, errors.Wrap(err, "wait for peers for data columns")
+	}
+
+	// Select the peers that will be requested.
+	dataColumnsToFetchByPeer, err := SelectPeersToFetchDataColumnsFrom(targetRequest.Columns, dataColumnsByAdmissiblePeer)
+	if err != nil {
+		// This should never happen.
+		return nil, errors.Wrap(err, "select peers to fetch data columns from")
+	}
+
+	var roDataColumns []blocks.RODataColumn
+	for peer, columnsToFetch := range dataColumnsToFetchByPeer {
+		// Build the request.
+		request := &eth.DataColumnSidecarsByRangeRequest{
+			StartSlot: targetRequest.StartSlot,
+			Count:     targetRequest.Count,
+			Columns:   columnsToFetch,
+		}
+
+		peerRoDataColumns, err := SendDataColumnSidecarsByRangeRequest(ctx, clock, p2p, peer, ctxMap, request)
+		if err != nil {
+			return nil, errors.Wrap(err, "send data column sidecars by range request")
+		}
+
+		roDataColumns = append(roDataColumns, peerRoDataColumns...)
+	}
+
+	return roDataColumns, nil
+}
+
+// waitForPeersForDataColumns returns a map, where the key of the map is the peer, the value is the custody columns of the peer.
+// It uses only peers
+// - synced up to `lastSlot`, and
+// - have bandwidth to serve `blockCount` blocks.
+// It waits until at least one peer per data column is available.
+func waitForPeersForDataColumns(p2p p2p.P2P, rateLimiter *leakybucket.Collector, peers []peer.ID, request *eth.DataColumnSidecarsByRangeRequest) (map[peer.ID]map[uint64]bool, error) {
+	const delay = 5 * time.Second
+
+	numberOfColumns := params.BeaconConfig().NumberOfColumns
+
+	// Build nice log fields.
+	lastSlot := request.StartSlot.Add(request.Count).Sub(1)
+
+	var neededDataColumnsLog interface{} = "all"
+	neededDataColumnCount := uint64(len(request.Columns))
+	if neededDataColumnCount < numberOfColumns {
+		neededDataColumnsLog = request.Columns
+	}
+
+	log := log.WithFields(logrus.Fields{
+		"start":             request.StartSlot,
+		"targetSlot":        lastSlot,
+		"neededDataColumns": neededDataColumnsLog,
+	})
+
+	// Keep only peers with head epoch greater than or equal to the epoch corresponding to the target slot, and
+	// keep only peers with enough bandwidth.
+	filteredPeers, descriptions, err := filterPeersByTargetSlotAndBandwidth(p2p, rateLimiter, peers, lastSlot, request.Count)
+	if err != nil {
+		return nil, errors.Wrap(err, "filter eers by target slot and bandwidth")
+	}
+
+	// Get the peers that are admissible for the data columns.
+	dataColumnsByAdmissiblePeer, admissiblePeersByDataColumn, moreDescriptions, err := AdmissiblePeersForDataColumns(filteredPeers, request.Columns, p2p)
+	if err != nil {
+		return nil, errors.Wrap(err, "admissible peers for data columns")
+	}
+
+	descriptions = append(descriptions, moreDescriptions...)
+
+	// Compute data columns without any peer.
+	dataColumnsWithoutPeers := computeDataColumnsWithoutPeers(request.Columns, admissiblePeersByDataColumn)
+
+	// Wait if no suitable peers are available.
+	for len(dataColumnsWithoutPeers) > 0 {
+		// Build a nice log fields.
+		var dataColumnsWithoutPeersLog interface{} = "all"
+		dataColumnsWithoutPeersCount := uint64(len(dataColumnsWithoutPeers))
+		if dataColumnsWithoutPeersCount < numberOfColumns {
+			dataColumnsWithoutPeersLog = uint64MapToSortedSlice(dataColumnsWithoutPeers)
+		}
+
+		log.WithField("columnsWithoutPeer", dataColumnsWithoutPeersLog).Warning("Fetch data columns from peers - no available peers, retrying later")
+		for _, description := range descriptions {
+			log.Debug(description)
+		}
+
+		for pid, peerDataColumns := range dataColumnsByAdmissiblePeer {
+			var peerDataColumnsLog interface{} = "all"
+			peerDataColumnsCount := uint64(len(peerDataColumns))
+			if peerDataColumnsCount < numberOfColumns {
+				peerDataColumnsLog = uint64MapToSortedSlice(peerDataColumns)
+			}
+
+			log.WithFields(logrus.Fields{
+				"peer":            pid,
+				"peerDataColumns": peerDataColumnsLog,
+			}).Debug("Peer data columns")
+		}
+
+		time.Sleep(delay)
+
+		// Filter for peers with head epoch greater than or equal to our target epoch for ByRange requests.
+		filteredPeers, descriptions, err = filterPeersByTargetSlotAndBandwidth(p2p, rateLimiter, peers, lastSlot, request.Count)
+		if err != nil {
+			return nil, errors.Wrap(err, "filter peers by target slot and bandwidth")
+		}
+
+		// Get the peers that are admissible for the data columns.
+		dataColumnsByAdmissiblePeer, admissiblePeersByDataColumn, moreDescriptions, err = AdmissiblePeersForDataColumns(filteredPeers, request.Columns, p2p)
+		if err != nil {
+			return nil, errors.Wrap(err, "admissible peers for data columns")
+		}
+
+		descriptions = append(descriptions, moreDescriptions...)
+
+		// Compute data columns without any peer.
+		dataColumnsWithoutPeers = computeDataColumnsWithoutPeers(request.Columns, admissiblePeersByDataColumn)
+	}
+
+	return dataColumnsByAdmissiblePeer, nil
+}
+
+// Filter peers to ensure they are synced to the target slot and have sufficient bandwidth to serve the request.
+func filterPeersByTargetSlotAndBandwidth(p2p p2p.P2P, rateLimiter *leakybucket.Collector, peers []peer.ID, lastSlot primitives.Slot, blockCount uint64) ([]peer.ID, []string, error) {
+	if len(peers) == 0 {
+		peers = p2p.Peers().Connected()
+	}
+
+	slotPeers, descriptions, err := filterPeersByTargetSlot(p2p, peers, lastSlot)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "peers with slot and data columns")
+	}
+
+	// Filter for peers with sufficient bandwidth to serve the request.
+	slotAndBandwidthPeers := hasSufficientBandwidth(rateLimiter, slotPeers, blockCount)
+
+	// Add debugging logs for the filtered peers.
+	peerWithSufficientBandwidthMap := make(map[peer.ID]bool, len(peers))
+	for _, peer := range slotAndBandwidthPeers {
+		peerWithSufficientBandwidthMap[peer] = true
+	}
+
+	for _, peer := range slotPeers {
+		if !peerWithSufficientBandwidthMap[peer] {
+			description := fmt.Sprintf("peer %s: does not have sufficient bandwidth", peer)
+			descriptions = append(descriptions, description)
+		}
+	}
+	return slotAndBandwidthPeers, descriptions, nil
+}
+
+func hasSufficientBandwidth(rateLimiter *leakybucket.Collector, peers []peer.ID, count uint64) []peer.ID {
+	var filteredPeers []peer.ID
+
+	for _, p := range peers {
+		if uint64(rateLimiter.Remaining(p.String())) < count {
+			continue
+		}
+		copiedP := p
+		filteredPeers = append(filteredPeers, copiedP)
+	}
+	return filteredPeers
+}
+
+func computeDataColumnsWithoutPeers(neededColumns []uint64, peersByColumn map[uint64][]peer.ID) map[uint64]bool {
+	result := make(map[uint64]bool)
+	for _, column := range neededColumns {
+		if _, ok := peersByColumn[column]; !ok {
+			result[column] = true
+		}
+	}
+
+	return result
+}
+
+// Filter peers with head epoch lower than our target epoch for ByRange requests.
+func filterPeersByTargetSlot(p2p p2p.P2P, peers []peer.ID, targetSlot primitives.Slot) ([]peer.ID, []string, error) {
+	filteredPeers := make([]peer.ID, 0, len(peers))
+	descriptions := make([]string, 0, len(peers))
+	// Compute the target epoch from the target slot.
+	targetEpoch := slots.ToEpoch(targetSlot)
+
+	for _, peer := range peers {
+		peerChainState, err := p2p.Peers().ChainState(peer)
+		if err != nil {
+			description := fmt.Sprintf("peer %s: error: %s", peer, err)
+			descriptions = append(descriptions, description)
+			continue
+		}
+
+		if peerChainState == nil {
+			description := fmt.Sprintf("peer %s: chain state is nil", peer)
+			descriptions = append(descriptions, description)
+			continue
+		}
+
+		peerHeadEpoch := slots.ToEpoch(peerChainState.HeadSlot)
+
+		if peerHeadEpoch < targetEpoch {
+			description := fmt.Sprintf("peer %s: peer head epoch %d < our target epoch %d", peer, peerHeadEpoch, targetEpoch)
+			descriptions = append(descriptions, description)
+			continue
+		}
+
+		filteredPeers = append(filteredPeers, peer)
+	}
+
+	return filteredPeers, descriptions, nil
+}
+
+// itemsCount returns the total count of items
+func itemsCount(missingColumnsByRoot map[[fieldparams.RootLength]byte]map[uint64]bool) int {
+	count := 0
+	for _, columns := range missingColumnsByRoot {
+		count += len(columns)
+	}
+	return count
+}
+
+// uint64MapDiffer returns true if the two maps differ.
+func uint64MapDiffer(left, right map[uint64]bool) bool {
+	if len(left) != len(right) {
+		return true
+	}
+
+	for k := range left {
+		if !right[k] {
+			return true
+		}
+	}
+
+	return false
 }
