+++ beacon-chain/sync/initial-sync/blocks_fetcher.go
@@ -20,7 +20,6 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/sync/verify"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/verification"
 	"github.com/prysmaticlabs/prysm/v5/cmd/beacon-chain/flags"
-	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
@@ -355,9 +354,15 @@ func (f *blocksFetcher) fetchSidecars(ctx context.Context, pid peer.ID, peers []
 	blocksWithBlobs := bwScs[:firstFuluIndex]
 	blocksWithDataColumns := bwScs[firstFuluIndex:]
 
-	// Fetch blob sidecars.
-	if err := f.fetchBlobsFromPeer(ctx, blocksWithBlobs, pid, peers); err != nil {
-		return errors.Wrap(err, "fetch blobs from peer")
+	if len(blocksWithBlobs) > 0 {
+		// Fetch blob sidecars.
+		if err := f.fetchBlobsFromPeer(ctx, blocksWithBlobs, pid, peers); err != nil {
+			return errors.Wrap(err, "fetch blobs from peer")
+		}
+	}
+
+	if len(blocksWithDataColumns) == 0 {
+		return nil
 	}
 
 	// Extract blocks.
@@ -368,7 +373,8 @@ func (f *blocksFetcher) fetchSidecars(ctx context.Context, pid peer.ID, peers []
 	}
 
 	// Fetch data column sidecars.
-	fetchedDataColumnsByRoot, err := f.fetchMissingDataColumnsFromPeers(ctx, dataColumnBlocks, peers, batchSize)
+	actualGroupCount := f.custodyInfo.ActualGroupCount()
+	fetchedDataColumnsByRoot, err := prysmsync.RequestMissingDataColumnsByRange(ctx, f.clock, f.ctxMap, f.p2p, f.rateLimiter, actualGroupCount, f.dcs, peers, dataColumnBlocks, batchSize)
 	if err != nil {
 		return errors.Wrap(err, "fetch missing data columns from peers")
 	}
@@ -657,418 +663,7 @@ func sortedSliceFromMap(m map[uint64]bool) []uint64 {
 	return result
 }
 
-// buildDataColumnByRangeRequests builds an optimized slices of data column by range requests:
-// 1. It will never request a data column that is already stored in the DB if there is no "hole" in `roBlocks` other than missed slots.
-// 2. It will minimize the number of requests.
-// It expects blocks to be sorted by slot.
-func buildDataColumnByRangeRequests(roBlocks []blocks.ROBlock, missingColumnsByRoot map[[fieldparams.RootLength]byte]map[uint64]bool, batchSize int) ([]*p2ppb.DataColumnSidecarsByRangeRequest, error) {
-	batchSizeSlot := primitives.Slot(batchSize)
-
-	// Return early if there are no blocks to process.
-	if len(roBlocks) == 0 {
-		return nil, nil
-	}
-
-	// It's safe to get the first item of the slice since we've already checked that it's not empty.
-	firstROBlock, lastROBlock := roBlocks[0], roBlocks[len(roBlocks)-1]
-	firstBlockSlot, lastBlockSlot := firstROBlock.Block().Slot(), lastROBlock.Block().Slot()
-	firstBlockRoot := firstROBlock.Root()
-
-	previousMissingDataColumns := make(map[uint64]bool, len(missingColumnsByRoot[firstBlockRoot]))
-
-	if missing, ok := missingColumnsByRoot[firstBlockRoot]; ok {
-		for key, value := range missing {
-			previousMissingDataColumns[key] = value
-		}
-	}
-
-	previousBlockSlot, previousStartBlockSlot := firstBlockSlot, firstBlockSlot
-
-	result := make([]*p2ppb.DataColumnSidecarsByRangeRequest, 0, 1)
-	for index := 1; index < len(roBlocks); index++ {
-		roBlock := roBlocks[index]
-
-		// Extract the block from the RO-block.
-		block := roBlock.Block()
-
-		// Extract the slot from the block.
-		blockRoot, blockSlot := roBlock.Root(), block.Slot()
-
-		if blockSlot <= previousBlockSlot {
-			return nil, errors.Errorf("blocks are not strictly sorted by slot. Previous block slot: %d, current block slot: %d", previousBlockSlot, blockSlot)
-		}
-
-		// Extract KZG commitments count from the current block body
-		blockKzgCommitments, err := block.Body().BlobKzgCommitments()
-		if err != nil {
-			return nil, errors.Wrap(err, "blob KZG commitments")
-		}
-
-		// Compute the count of KZG commitments.
-		blockKzgCommitmentCount := len(blockKzgCommitments)
-
-		// Skip blocks without commitments.
-		if blockKzgCommitmentCount == 0 {
-			previousBlockSlot = blockSlot
-			continue
-		}
-
-		// Get the missing data columns for the current block.
-		missingDataColumns := make(map[uint64]bool, len(missingColumnsByRoot[blockRoot]))
-		for key, value := range missingColumnsByRoot[blockRoot] {
-			missingDataColumns[key] = value
-		}
-
-		// Compute if the missing data columns differ.
-		missingDataColumnsDiffer := uint64MapDiffer(previousMissingDataColumns, missingDataColumns)
-
-		// Compute if the batch size is reached.
-		batchSizeReached := blockSlot-previousStartBlockSlot >= batchSizeSlot
-
-		if missingDataColumnsDiffer || batchSizeReached {
-			// Append the slice to the result.
-			request := &p2ppb.DataColumnSidecarsByRangeRequest{
-				StartSlot: previousStartBlockSlot,
-				Count:     uint64(blockSlot - previousStartBlockSlot),
-				Columns:   sortedSliceFromMap(previousMissingDataColumns),
-			}
-
-			result = append(result, request)
-
-			previousStartBlockSlot, previousMissingDataColumns = blockSlot, missingDataColumns
-		}
-
-		previousBlockSlot = blockSlot
-	}
-
-	lastRequest := &p2ppb.DataColumnSidecarsByRangeRequest{
-		StartSlot: previousStartBlockSlot,
-		Count:     uint64(lastBlockSlot - previousStartBlockSlot + 1),
-		Columns:   sortedSliceFromMap(previousMissingDataColumns),
-	}
-
-	result = append(result, lastRequest)
-
-	return result, nil
-}
-
-// uint64MapDiffer returns true if the two maps differ.
-func uint64MapDiffer(left, right map[uint64]bool) bool {
-	if len(left) != len(right) {
-		return true
-	}
-
-	for k := range left {
-		if !right[k] {
-			return true
-		}
-	}
-
-	return false
-}
-
-// computeMissingDataColumnsCount returns the count of missing columns.
-func computeMissingDataColumnsCount(missingColumnsByRoot map[[fieldparams.RootLength]byte]map[uint64]bool) int {
-	count := 0
-	for _, columns := range missingColumnsByRoot {
-		count += len(columns)
-	}
-	return count
-}
-
-// fetchDataColumnsFromPeers requests data columns by range to relevant peers
-func (f *blocksFetcher) fetchDataColumnsFromPeers(ctx context.Context, peers []peer.ID, targetRequest *p2ppb.DataColumnSidecarsByRangeRequest) ([]blocks.RODataColumn, error) {
-	// Filter out requests with no data columns.
-	if len(targetRequest.Columns) == 0 {
-		return nil, nil
-	}
-
-	// Get all admissible peers with the data columns they custody.
-	dataColumnsByAdmissiblePeer, err := f.waitForPeersForDataColumns(peers, targetRequest)
-	if err != nil {
-		return nil, errors.Wrap(err, "wait for peers for data columns")
-	}
-
-	// Select the peers that will be requested.
-	dataColumnsToFetchByPeer, err := prysmsync.SelectPeersToFetchDataColumnsFrom(targetRequest.Columns, dataColumnsByAdmissiblePeer)
-	if err != nil {
-		// This should never happen.
-		return nil, errors.Wrap(err, "select peers to fetch data columns from")
-	}
-
-	var roDataColumns []blocks.RODataColumn
-	for peer, columnsToFetch := range dataColumnsToFetchByPeer {
-		// Build the request.
-		request := &p2ppb.DataColumnSidecarsByRangeRequest{
-			StartSlot: targetRequest.StartSlot,
-			Count:     targetRequest.Count,
-			Columns:   columnsToFetch,
-		}
-
-		peerRoDataColumns, err := prysmsync.SendDataColumnSidecarsByRangeRequest(ctx, f.clock, f.p2p, peer, f.ctxMap, request)
-		if err != nil {
-			return nil, errors.Wrap(err, "send data column sidecars by range request")
-		}
-
-		roDataColumns = append(roDataColumns, peerRoDataColumns...)
-	}
-
-	return roDataColumns, nil
-}
-
-// fetchMissingDataColumnsFromPeers looks at the blocks with given roots and retrieves all
-// data columns for blocks that have commitments, and for which our store is missing data columns we should custody.
-// It expects blocks to be sorted by slot.
-func (f *blocksFetcher) fetchMissingDataColumnsFromPeers(ctx context.Context, blks []blocks.ROBlock, peers []peer.ID, batchSize int) (map[[fieldparams.RootLength]byte][]blocks.RODataColumn, error) {
-	const maxAllowedStall = 5 // Number of trials before giving up.
-
-	if len(blks) == 0 {
-		return nil, nil
-	}
-
-	// Compute the current slot.
-	currentSlot := f.clock.CurrentSlot()
-
-	// Compute the minimum slot for which we should serve data columns.
-	minimumSlot, err := prysmsync.DataColumnsRPCMinValidSlot(currentSlot)
-	if err != nil {
-		return nil, errors.Wrap(err, "data columns RPC min valid slot")
-	}
-
-	// Get the node ID.
-	nodeID := f.p2p.NodeID()
-
-	// Get the actual group count.
-	actualGroupCount := f.custodyInfo.ActualGroupCount()
-
-	// Get blocks by root and compute all missing columns by root.
-	blockByRoot := make(map[[fieldparams.RootLength]byte]blocks.ROBlock, len(blks))
-	missingColumnsByRoot := make(map[[fieldparams.RootLength]byte]map[uint64]bool, len(blks))
-	for _, blk := range blks {
-		// Extract the block root and the block slot
-		blockRoot, blockSlot := blk.Root(), blk.Block().Slot()
-
-		// Populate the block by root.
-		blockByRoot[blockRoot] = blk
-
-		// Skip blocks that are not in the retention period.
-		if blockSlot < minimumSlot {
-			continue
-		}
-
-		missingColumns, err := prysmsync.MissingDataColumns(blk, nodeID, actualGroupCount, f.dcs)
-		if err != nil {
-			return nil, errors.Wrap(err, "missing data columns")
-		}
-
-		for _, column := range missingColumns {
-			if _, ok := missingColumnsByRoot[blockRoot]; !ok {
-				missingColumnsByRoot[blockRoot] = make(map[uint64]bool)
-			}
-			missingColumnsByRoot[blockRoot][column] = true
-		}
-	}
-
-	// Return early if there are no missing data columns.
-	if len(missingColumnsByRoot) == 0 {
-		return nil, nil
-	}
-
-	// Compute the number of missing data columns.
-	previousMissingDataColumnsCount := computeMissingDataColumnsCount(missingColumnsByRoot)
-
-	// Count the number of retries for the same amount of missing data columns.
-	stallCount := 0
-
-	// Add log fields.
-	log := log.WithFields(logrus.Fields{
-		"initialMissingColumnsCount": previousMissingDataColumnsCount,
-		"blockCount":                 len(blks),
-		"firstSlot":                  blks[0].Block().Slot(),
-		"lastSlot":                   blks[len(blks)-1].Block().Slot(),
-	})
-
-	// Log the start of the process.
-	start := time.Now()
-	log.Debug("Requesting data column sidecars - start")
-
-	alignedDataColumnsByRoot := make(map[[fieldparams.RootLength]byte][]blocks.RODataColumn, len(blks))
-	for len(missingColumnsByRoot) > 0 {
-		// Build requests.
-		requests, err := buildDataColumnByRangeRequests(blks, missingColumnsByRoot, batchSize)
-		if err != nil {
-			return nil, errors.Wrap(err, "build data column by range requests")
-		}
-
-		// Requests data column sidecars from peers.
-		retrievedDataColumnsByRoot := make(map[[fieldparams.RootLength]byte][]blocks.RODataColumn)
-		for _, request := range requests {
-			roDataColumns, err := f.fetchDataColumnsFromPeers(ctx, peers, request)
-			if err != nil {
-				return nil, errors.Wrap(err, "fetch data columns from peers")
-			}
-
-			for _, roDataColumn := range roDataColumns {
-				root := roDataColumn.BlockRoot()
-				if _, ok := blockByRoot[root]; !ok {
-					// It may happen if the peer which sent the data columns is on a different fork.
-					continue
-				}
-
-				retrievedDataColumnsByRoot[root] = append(retrievedDataColumnsByRoot[root], roDataColumn)
-			}
-		}
-
-		for root, dataColumns := range retrievedDataColumnsByRoot {
-			// Retrieve the block from the root.
-			block, ok := blockByRoot[root]
-			if !ok {
-				return nil, errors.New("block not found - this should never happen")
-			}
-
-			// Check if the data columns align with blocks.
-			if err := verify.DataColumnsAlignWithBlock(block, dataColumns); err != nil {
-				log.WithField("root", root).WithError(err).Debug("Data columns do not align with block")
-				continue
-			}
-
-			alignedDataColumnsByRoot[root] = append(alignedDataColumnsByRoot[root], dataColumns...)
-
-			// Remove aligned data columns from the missing columns.
-			for _, dataColumn := range dataColumns {
-				delete(missingColumnsByRoot[root], dataColumn.Index)
-				if len(missingColumnsByRoot[root]) == 0 {
-					delete(missingColumnsByRoot, root)
-				}
-			}
-		}
-
-		missingDataColumnsCount := computeMissingDataColumnsCount(missingColumnsByRoot)
-		if missingDataColumnsCount == previousMissingDataColumnsCount {
-			stallCount++
-		} else {
-			stallCount = 0
-		}
-
-		previousMissingDataColumnsCount = missingDataColumnsCount
-
-		if missingDataColumnsCount > 0 {
-			log := log.WithFields(logrus.Fields{
-				"remainingMissingColumnsCount": missingDataColumnsCount,
-				"stallCount":                   stallCount,
-				"maxAllowedStall":              maxAllowedStall,
-			})
-
-			if stallCount >= maxAllowedStall {
-				// It is very likely `bwbs` contains orphaned blocks, for which no peer has the data columns.
-				// We give up and let the state machine handle the situation.
-				const message = "Requesting data column sidecars - no progress, giving up"
-				log.Warning(message)
-				return nil, errors.New(message)
-			}
-
-			log.WithFields(logrus.Fields{
-				"remainingMissingColumnsCount": missingDataColumnsCount,
-				"stallCount":                   stallCount,
-			}).Debug("Requesting data column sidecars - continue")
-		}
-	}
-
-	log.WithField("duration", time.Since(start)).Debug("Requesting data column sidecars - success")
-	return alignedDataColumnsByRoot, nil
-}
-
-// waitForPeersForDataColumns returns a map, where the key of the map is the peer, the value is the custody columns of the peer.
-// It uses only peers
-// - synced up to `lastSlot`, and
-// - have bandwidth to serve `blockCount` blocks.
-// It waits until at least one peer per data column is available.
-func (f *blocksFetcher) waitForPeersForDataColumns(peers []peer.ID, request *p2ppb.DataColumnSidecarsByRangeRequest) (map[peer.ID]map[uint64]bool, error) {
-	const delay = 5 * time.Second
-
-	// Keep only peers with head epoch greater than or equal to the epoch corresponding to the target slot, and
-	// keep only peers with enough bandwidth.
-	lastSlot := request.StartSlot.Add(request.Count).Sub(1)
-	filteredPeers, descriptions, err := f.filterPeersByTargetSlotAndBandwidth(peers, lastSlot, request.Count)
-	if err != nil {
-		return nil, errors.Wrap(err, "filter eers by target slot and bandwidth")
-	}
-
-	// Get the peers that are admissible for the data columns.
-	dataColumnsByAdmissiblePeer, admissiblePeersByDataColumn, moreDescriptions, err := prysmsync.AdmissiblePeersForDataColumns(filteredPeers, request.Columns, f.p2p)
-	if err != nil {
-		return nil, errors.Wrap(err, "admissible peers for data columns")
-	}
-
-	descriptions = append(descriptions, moreDescriptions...)
-
-	// Compute data columns without any peer.
-	dataColumnsWithoutPeers := computeDataColumnsWithoutPeers(request.Columns, admissiblePeersByDataColumn)
-
-	// Wait if no suitable peers are available.
-	for len(dataColumnsWithoutPeers) > 0 {
-		// Build a nice log fields.
-		numberOfColumns := params.BeaconConfig().NumberOfColumns
-
-		var neededDataColumnsLog interface{} = "all"
-		neededDataColumnCount := uint64(len(request.Columns))
-		if neededDataColumnCount < numberOfColumns {
-			neededDataColumnsLog = request.Columns
-		}
-
-		var dataColumnsWithoutPeersLog interface{} = "all"
-		dataColumnsWithoutPeersCount := uint64(len(dataColumnsWithoutPeers))
-		if dataColumnsWithoutPeersCount < numberOfColumns {
-			dataColumnsWithoutPeersLog = uint64MapToSortedSlice(dataColumnsWithoutPeers)
-		}
-
-		log.
-			WithFields(logrus.Fields{
-				"targetSlot":          lastSlot,
-				"neededDataColumns":   neededDataColumnsLog,
-				"columnsWithoutPeers": dataColumnsWithoutPeersLog,
-			}).
-			Warning("Fetch data columns from peers - no available peers, retrying later")
-
-		for _, description := range descriptions {
-			log.Debug(description)
-		}
-
-		for pid, peerDataColumns := range dataColumnsByAdmissiblePeer {
-			var peerDataColumnsLog interface{} = "all"
-			peerDataColumnsCount := uint64(len(peerDataColumns))
-			if peerDataColumnsCount < numberOfColumns {
-				peerDataColumnsLog = uint64MapToSortedSlice(peerDataColumns)
-			}
-
-			log.WithFields(logrus.Fields{
-				"peer":            pid,
-				"peerDataColumns": peerDataColumnsLog,
-			}).Debug("Peer data columns")
-		}
-
-		time.Sleep(delay)
-
-		// Filter for peers with head epoch greater than or equal to our target epoch for ByRange requests.
-		filteredPeers, descriptions, err = f.filterPeersByTargetSlotAndBandwidth(peers, lastSlot, request.Count)
-		if err != nil {
-			return nil, errors.Wrap(err, "filter peers by target slot and bandwidth")
-		}
-
-		// Get the peers that are admissible for the data columns.
-		dataColumnsByAdmissiblePeer, admissiblePeersByDataColumn, moreDescriptions, err = prysmsync.AdmissiblePeersForDataColumns(filteredPeers, request.Columns, f.p2p)
-		if err != nil {
-			return nil, errors.Wrap(err, "admissible peers for data columns")
-		}
-
-		descriptions = append(descriptions, moreDescriptions...)
-
-		// Compute data columns without any peer.
-		dataColumnsWithoutPeers = computeDataColumnsWithoutPeers(request.Columns, admissiblePeersByDataColumn)
-	}
-
-	return dataColumnsByAdmissiblePeer, nil
-}
+// waitForPeersFo
 
 // requestBlocks is a wrapper for handling BeaconBlocksByRangeRequest requests/streams.
 func (f *blocksFetcher) requestBlocks(
@@ -1228,15 +823,4 @@ func dedupPeers(peers []peer.ID) []peer.ID {
 		peerExists[peers[i]] = true
 	}
 	return newPeerList
-}
-
-func computeDataColumnsWithoutPeers(neededColumns []uint64, peersByColumn map[uint64][]peer.ID) map[uint64]bool {
-	result := make(map[uint64]bool)
-	for _, column := range neededColumns {
-		if _, ok := peersByColumn[column]; !ok {
-			result[column] = true
-		}
-	}
-
-	return result
 }
