+++ beacon-chain/sync/initial-sync/blocks_fetcher_test.go
@@ -1,11 +1,7 @@
 package initialsync
 
 import (
-	"bytes"
 	"context"
-	"crypto/sha256"
-	"encoding/binary"
-	"encoding/hex"
 	"fmt"
 	"math"
 	"sort"
@@ -13,27 +9,17 @@ import (
 	"testing"
 	"time"
 
-	"github.com/consensys/gnark-crypto/ecc/bls12-381/fr"
-	GoKZG "github.com/crate-crypto/go-kzg-4844"
-	"github.com/ethereum/go-ethereum/p2p/enr"
-	"github.com/libp2p/go-libp2p"
 	libp2pcore "github.com/libp2p/go-libp2p/core"
-	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/network"
 	"github.com/libp2p/go-libp2p/core/peer"
-	"github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain/kzg"
 	mock "github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain/testing"
-	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/peerdas"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/db/filesystem"
 	dbtest "github.com/prysmaticlabs/prysm/v5/beacon-chain/db/testing"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/p2p"
-	"github.com/prysmaticlabs/prysm/v5/beacon-chain/p2p/peers"
 	p2ptest "github.com/prysmaticlabs/prysm/v5/beacon-chain/p2p/testing"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/startup"
 	beaconsync "github.com/prysmaticlabs/prysm/v5/beacon-chain/sync"
-	"github.com/prysmaticlabs/prysm/v5/beacon-chain/verification"
 	"github.com/prysmaticlabs/prysm/v5/cmd/beacon-chain/flags"
-	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
@@ -41,9 +27,7 @@ import (
 	leakybucket "github.com/prysmaticlabs/prysm/v5/container/leaky-bucket"
 	"github.com/prysmaticlabs/prysm/v5/container/slice"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
-	"github.com/prysmaticlabs/prysm/v5/network/forks"
 	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
-	"github.com/prysmaticlabs/prysm/v5/runtime/version"
 	"github.com/prysmaticlabs/prysm/v5/testing/assert"
 	"github.com/prysmaticlabs/prysm/v5/testing/require"
 	"github.com/prysmaticlabs/prysm/v5/testing/util"
@@ -1320,914 +1304,4 @@ func TestSortedSliceFromMap(t *testing.T) {
 
 	actual := sortedSliceFromMap(m)
 	require.DeepSSZEqual(t, expected, actual)
-}
-
-type blockParams struct {
-	slot     primitives.Slot
-	hasBlobs bool
-}
-
-func deterministicRandomness(t *testing.T, seed int64) [32]byte {
-	buf := new(bytes.Buffer)
-	err := binary.Write(buf, binary.BigEndian, seed)
-	require.NoError(t, err)
-	bytes := buf.Bytes()
-
-	return sha256.Sum256(bytes)
-}
-
-// getRandFieldElement returns a serialized random field element in big-endian
-func getRandFieldElement(t *testing.T, seed int64) [32]byte {
-	bytes := deterministicRandomness(t, seed)
-	var r fr.Element
-	r.SetBytes(bytes[:])
-
-	return GoKZG.SerializeScalar(r)
-}
-
-// getRandBlob returns a random blob using the passed seed as entropy
-func getRandBlob(t *testing.T, seed int64) kzg.Blob {
-	var blob kzg.Blob
-	for i := 0; i < len(blob); i += 32 {
-		fieldElementBytes := getRandFieldElement(t, seed+int64(i))
-		copy(blob[i:i+32], fieldElementBytes[:])
-	}
-	return blob
-}
-
-type (
-	responseParams struct {
-		slot        primitives.Slot
-		columnIndex uint64
-	}
-
-	peerParams struct {
-		// Custody subnet count
-		cgc uint64
-
-		// key: RPCDataColumnSidecarsByRangeTopicV1 stringified
-		// value: The list of all slotxindex to respond by request number
-		toRespond map[string][][]responseParams
-	}
-)
-
-// createAndConnectPeer creates a peer and connects it to the p2p service.
-// The peer will respond to the `RPCDataColumnSidecarsByRangeTopicV1` topic.
-func createAndConnectPeer(
-	t *testing.T,
-	p2pService *p2ptest.TestP2P,
-	chainService *mock.ChainService,
-	dataColumnsSidecarFromSlot map[primitives.Slot][]*ethpb.DataColumnSidecar,
-	peerParams peerParams,
-	offset int,
-) *p2ptest.TestP2P {
-	// Create the private key, depending on the offset.
-	privateKeyBytes := make([]byte, 32)
-	for i := range 32 {
-		privateKeyBytes[i] = byte(offset + i)
-	}
-
-	privateKey, err := crypto.UnmarshalSecp256k1PrivateKey(privateKeyBytes)
-	require.NoError(t, err)
-
-	// Create the peer.
-	peer := p2ptest.NewTestP2P(t, libp2p.Identity(privateKey))
-
-	// Create a call counter.
-	countFromRequest := make(map[string]int, len(peerParams.toRespond))
-
-	peer.SetStreamHandler(p2p.RPCDataColumnSidecarsByRangeTopicV1+"/ssz_snappy", func(stream network.Stream) {
-		// Decode the request.
-		req := new(ethpb.DataColumnSidecarsByRangeRequest)
-
-		err := peer.Encoding().DecodeWithMaxLength(stream, req)
-		require.NoError(t, err)
-
-		// Convert the request to a string.
-		reqString := req.String()
-
-		// Get the response to send.
-		items, ok := peerParams.toRespond[reqString]
-		require.Equal(t, true, ok, "no response to send for request %s", reqString)
-
-		for _, responseParams := range items[countFromRequest[reqString]] {
-			// Get data columns sidecars for this slot.
-			dataColumnsSidecar, ok := dataColumnsSidecarFromSlot[responseParams.slot]
-			require.Equal(t, true, ok)
-
-			// Get the data column sidecar.
-			dataColumn := dataColumnsSidecar[responseParams.columnIndex]
-
-			// Send the response.
-			err := beaconsync.WriteDataColumnSidecarChunk(stream, chainService, p2pService.Encoding(), dataColumn)
-			require.NoError(t, err)
-		}
-
-		// Close the stream.
-		err = stream.Close()
-		require.NoError(t, err)
-
-		// Increment the call counter.
-		countFromRequest[reqString]++
-	})
-
-	// Create the record and set the custody count.
-	enr := &enr.Record{}
-	enr.Set(peerdas.Cgc(peerParams.cgc))
-
-	// Add the peer and connect it.
-	p2pService.Peers().Add(enr, peer.PeerID(), nil, network.DirOutbound)
-	p2pService.Peers().SetConnectionState(peer.PeerID(), peers.Connected)
-	p2pService.Connect(peer)
-
-	return peer
-}
-
-func defaultMockChain(t *testing.T, currentSlot uint64) (*mock.ChainService, *startup.Clock) {
-	de := params.BeaconConfig().FuluForkEpoch
-	df, err := forks.Fork(de)
-	require.NoError(t, err)
-	denebBuffer := params.BeaconConfig().MinEpochsForBlobsSidecarsRequest + 1000
-	ce := de + denebBuffer
-	fe := ce - 2
-	cs, err := slots.EpochStart(ce)
-	require.NoError(t, err)
-	now := time.Now()
-	genOffset := primitives.Slot(params.BeaconConfig().SecondsPerSlot) * cs
-	genesisTime := now.Add(-1 * time.Second * time.Duration(int64(genOffset)))
-
-	clock := startup.NewClock(genesisTime, [32]byte{}, startup.WithNower(
-		func() time.Time {
-			return genesisTime.Add(time.Duration(currentSlot*params.BeaconConfig().SecondsPerSlot) * time.Second)
-		},
-	))
-
-	chain := &mock.ChainService{
-		FinalizedCheckPoint: &ethpb.Checkpoint{Epoch: fe},
-		Fork:                df,
-	}
-
-	return chain, clock
-}
-
-func TestBuildDataColumnByRangeRequests(t *testing.T) {
-	type missingColumnsWithCommitment struct {
-		areCommitments bool
-		missingColumns map[uint64]bool
-	}
-
-	testCases := []struct {
-		name      string
-		batchSize int
-
-		// input
-		missingColumnsWithCommitments []*missingColumnsWithCommitment
-
-		// output
-		requests []*ethpb.DataColumnSidecarsByRangeRequest
-	}{
-		{
-			name:                          "no item",
-			batchSize:                     32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{},
-			requests:                      nil,
-		},
-		{
-			name:                          "one item - no missing columns",
-			batchSize:                     32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{{areCommitments: true, missingColumns: map[uint64]bool{}}},
-			requests:                      []*ethpb.DataColumnSidecarsByRangeRequest{{StartSlot: 0, Count: 1, Columns: []uint64{}}},
-		},
-		{
-			name:                          "one item - some missing columns",
-			batchSize:                     32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}}},
-			requests:                      []*ethpb.DataColumnSidecarsByRangeRequest{{StartSlot: 0, Count: 1, Columns: []uint64{1, 3, 5}}},
-		},
-		{
-			name:      "two items - no break",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{{StartSlot: 0, Count: 2, Columns: []uint64{1, 3, 5}}},
-		},
-		{
-			name:      "three items - no break",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{{StartSlot: 0, Count: 3, Columns: []uint64{1, 3, 5}}},
-		},
-		{
-			name:      "five items - columns break",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 2, Columns: []uint64{1, 3, 5}},
-				{StartSlot: 2, Count: 2, Columns: []uint64{1, 3}},
-				{StartSlot: 4, Count: 1, Columns: []uint64{}},
-			},
-		},
-		{
-			name:      "seven items - gap",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				nil,
-				nil,
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 7, Columns: []uint64{1, 3, 5}},
-			},
-		},
-		{
-			name:      "seven items - only breaks",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				nil,
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{2: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 1, Columns: []uint64{}},
-				{StartSlot: 1, Count: 3, Columns: []uint64{1, 3, 5}},
-				{StartSlot: 4, Count: 1, Columns: []uint64{2}},
-				{StartSlot: 5, Count: 1, Columns: []uint64{}},
-			},
-		},
-		{
-			name:      "thirteen items - some blocks without commitments",
-			batchSize: 32,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				nil,
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{2: true, 4: true}},
-				{areCommitments: false, missingColumns: nil},
-				{areCommitments: false, missingColumns: nil},
-				{areCommitments: true, missingColumns: map[uint64]bool{2: true, 4: true}},
-				nil,
-				nil,
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true}},
-				{areCommitments: false, missingColumns: nil},
-				{areCommitments: false, missingColumns: nil},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 4, Columns: []uint64{1, 3, 5}},
-				{StartSlot: 4, Count: 6, Columns: []uint64{2, 4}},
-				{StartSlot: 10, Count: 4, Columns: []uint64{1}},
-			},
-		},
-		{
-			name:      "five items - no break, limiting batch size",
-			batchSize: 3,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 3, Columns: []uint64{1, 3, 5}},
-				{StartSlot: 3, Count: 2, Columns: []uint64{1, 3, 5}},
-			},
-		},
-		{
-			name:      "eleven items - columns break, limiting batch size",
-			batchSize: 3,
-			missingColumnsWithCommitments: []*missingColumnsWithCommitment{
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true, 5: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{1: true, 3: true}},
-				{areCommitments: true, missingColumns: map[uint64]bool{}},
-				{areCommitments: false, missingColumns: nil},
-				{areCommitments: false, missingColumns: nil},
-				{areCommitments: true, missingColumns: map[uint64]bool{}},
-			},
-			requests: []*ethpb.DataColumnSidecarsByRangeRequest{
-				{StartSlot: 0, Count: 3, Columns: []uint64{1, 3, 5}},
-				{StartSlot: 3, Count: 3, Columns: []uint64{1, 3}},
-				{StartSlot: 6, Count: 1, Columns: []uint64{1, 3}},
-				{StartSlot: 7, Count: 3, Columns: []uint64{}},
-				{StartSlot: 10, Count: 1, Columns: []uint64{}},
-			},
-		},
-	}
-
-	// We don't care about the actual content of commitments, so we use a fake commitment.
-	fakeCommitment := make([]byte, 48)
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			roBlocks := make([]blocks.ROBlock, 0, len(tt.missingColumnsWithCommitments))
-			missingColumnsByRoot := make(map[[fieldparams.RootLength]byte]map[uint64]bool, len(tt.missingColumnsWithCommitments))
-			for i, missingColumnsWithCommitments := range tt.missingColumnsWithCommitments {
-				if missingColumnsWithCommitments == nil {
-					continue
-				}
-
-				missingColumns := missingColumnsWithCommitments.missingColumns
-
-				pbSignedBeaconBlock := util.NewBeaconBlockFulu()
-
-				signedBeaconBlock, err := blocks.NewSignedBeaconBlock(pbSignedBeaconBlock)
-				require.NoError(t, err)
-
-				signedBeaconBlock.SetSlot(primitives.Slot(i))
-
-				if missingColumnsWithCommitments.areCommitments {
-					err := signedBeaconBlock.SetBlobKzgCommitments([][]byte{fakeCommitment})
-					require.NoError(t, err)
-				}
-
-				roBlock, err := blocks.NewROBlock(signedBeaconBlock)
-				require.NoError(t, err)
-
-				roBlocks = append(roBlocks, roBlock)
-				roBlockRoot := roBlock.Root()
-
-				missingColumnsByRoot[roBlockRoot] = missingColumns
-			}
-
-			requests, err := buildDataColumnByRangeRequests(roBlocks, missingColumnsByRoot, tt.batchSize)
-			require.NoError(t, err)
-			require.DeepSSZEqual(t, tt.requests, requests)
-		})
-	}
-}
-
-func TestFetchDataColumnsFromPeers(t *testing.T) {
-	const (
-		blobsCount    = 6
-		peersHeadSlot = 100
-	)
-
-	testCases := []struct {
-		// Name of the test case.
-		name string
-
-		// INPUTS
-		// ------
-
-		// Fork epochs.
-		fuluForkEpoch primitives.Epoch
-
-		// Current slot.
-		currentSlot uint64
-
-		// Blocks with blobs parameters that will be used as `bwb` parameter.
-		blocksParams []blockParams
-
-		// What data columns do we store for the block in the same position in blocksParams.
-		// len(storedDataColumns) has to be the same than len(blocksParams).
-		storedDataColumns []map[uint64]bool
-
-		// Each item in the list represents a peer.
-		// We can specify what the peer will respond to each data column by range request.
-		// For the exact same data columns by range request, the peer will respond in the order they are specified.
-		peersParams []peerParams
-
-		// The max count of data columns that will be requested in each batch.
-		batchSize int
-
-		// OUTPUTS
-		// -------
-
-		// Data columns that should be added to `bwb`.
-		addedRODataColumns [][]int
-		isError            bool
-	}{
-		{
-			name:          "Fulu fork epoch is set to far futur epoch",
-			fuluForkEpoch: primitives.Epoch(math.MaxUint64),
-			blocksParams: []blockParams{
-				{slot: 1, hasBlobs: true}, // Before Fulu fork epoch
-				{slot: 2, hasBlobs: true}, // Before Fulu fork epoch
-				{slot: 3, hasBlobs: true}, // Before Fulu fork epoch
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{nil, nil, nil},
-			isError:            false,
-		},
-		{
-			name:          "All blocks are before Fulu fork epoch",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 25, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 26, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 27, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 28, hasBlobs: false}, // Before Fulu fork epoch
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{nil, nil, nil, nil},
-			isError:            false,
-		},
-		{
-			name:          "All blocks with commitments are before Fulu fork epoch",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 25, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 26, hasBlobs: true},  // Before Fulu fork epoch
-				{slot: 27, hasBlobs: true},  // Before Fulu fork epoch
-				{slot: 32, hasBlobs: false},
-				{slot: 33, hasBlobs: false},
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{nil, nil, nil, nil, nil},
-		},
-		{
-			name:          "Some blocks with blobs but without any missing data columns",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 25, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 26, hasBlobs: true},  // Before Fulu fork epoch
-				{slot: 27, hasBlobs: true},  // Before Fulu fork epoch
-				{slot: 32, hasBlobs: false},
-				{slot: 33, hasBlobs: true},
-			},
-			storedDataColumns: []map[uint64]bool{
-				nil,
-				nil,
-				nil,
-				nil,
-				{6: true, 38: true, 70: true, 102: true},
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{nil, nil, nil, nil, nil},
-			isError:            false,
-		},
-		{
-			name:          "Some blocks with blobs with missing data columns - one round needed",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 25, hasBlobs: false}, // Before Fulu fork epoch
-				{slot: 27, hasBlobs: true},  // Before Fulu fork epoch
-				{slot: 32, hasBlobs: false},
-				{slot: 33, hasBlobs: true},
-				{slot: 34, hasBlobs: true},
-				{slot: 35, hasBlobs: false},
-				{slot: 36, hasBlobs: true},
-				{slot: 37, hasBlobs: true},
-				{slot: 38, hasBlobs: true},
-				{slot: 39, hasBlobs: false},
-			},
-			storedDataColumns: []map[uint64]bool{
-				nil,                                      // Slot 25
-				nil,                                      // Slot 27
-				nil,                                      // Slot 32
-				{6: true, 38: true},                      // Slot 33
-				{6: true, 38: true},                      // Slot 34
-				nil,                                      // Slot 35
-				{6: true, 38: true},                      // Slot 36
-				{38: true, 102: true},                    // Slot 37
-				{6: true, 38: true, 70: true, 102: true}, // Slot 38
-				nil,                                      // Slot 39
-			},
-			peersParams: []peerParams{
-				{
-					cgc:       0,
-					toRespond: map[string][][]responseParams{},
-				},
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 33,
-							Count:     4,
-							Columns:   []uint64{70, 102},
-						}).String(): {
-							{
-								{slot: 33, columnIndex: 70},
-								{slot: 33, columnIndex: 102},
-								{slot: 34, columnIndex: 70},
-								{slot: 34, columnIndex: 102},
-								{slot: 36, columnIndex: 70},
-								{slot: 36, columnIndex: 102},
-							},
-						},
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 37,
-							Count:     1,
-							Columns:   []uint64{6, 70},
-						}).String(): {
-							{
-								{slot: 37, columnIndex: 6},
-								{slot: 37, columnIndex: 70},
-							},
-						},
-					},
-				},
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 33,
-							Count:     4,
-							Columns:   []uint64{70, 102},
-						}).String(): {
-							{
-								{slot: 33, columnIndex: 70},
-								{slot: 33, columnIndex: 102},
-								{slot: 34, columnIndex: 70},
-								{slot: 34, columnIndex: 102},
-								{slot: 36, columnIndex: 70},
-								{slot: 36, columnIndex: 102},
-							},
-						},
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 37,
-							Count:     1,
-							Columns:   []uint64{6, 70},
-						}).String(): {
-							{
-								{slot: 37, columnIndex: 6},
-								{slot: 37, columnIndex: 70},
-							},
-						},
-					},
-				},
-			},
-			batchSize: 32,
-			addedRODataColumns: [][]int{
-				nil,       // Slot 25
-				nil,       // Slot 27
-				nil,       // Slot 32
-				{70, 102}, // Slot 33
-				{70, 102}, // Slot 34
-				nil,       // Slot 35
-				{70, 102}, // Slot 36
-				{6, 70},   // Slot 37
-				nil,       // Slot 38
-				nil,       // Slot 39
-			},
-			isError: false,
-		},
-		{
-			name:          "Some blocks with blobs with missing data columns - partial responses",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 33, hasBlobs: true},
-				{slot: 34, hasBlobs: true},
-				{slot: 35, hasBlobs: false},
-				{slot: 36, hasBlobs: true},
-			},
-			storedDataColumns: []map[uint64]bool{
-				{6: true, 38: true}, // Slot 33
-				{6: true, 38: true}, // Slot 34
-				nil,                 // Slot 35
-				{6: true, 38: true}, // Slot 36
-			},
-			peersParams: []peerParams{
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 33,
-							Count:     4,
-							Columns:   []uint64{70, 102},
-						}).String(): {
-							{
-								{slot: 33, columnIndex: 70},
-								{slot: 34, columnIndex: 70},
-								{slot: 36, columnIndex: 70},
-							},
-						},
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 33,
-							Count:     4,
-							Columns:   []uint64{102},
-						}).String(): {
-							{
-								{slot: 33, columnIndex: 102},
-								{slot: 34, columnIndex: 102},
-								{slot: 36, columnIndex: 102},
-							},
-						},
-					},
-				},
-			},
-			batchSize: 32,
-			addedRODataColumns: [][]int{
-				{70, 102}, // Slot 33
-				{70, 102}, // Slot 34
-				nil,       // Slot 35
-				{70, 102}, // Slot 36
-			},
-		},
-		{
-			name:              "Some blocks with blobs with missing data columns - first response is empty",
-			fuluForkEpoch:     1,
-			currentSlot:       40,
-			blocksParams:      []blockParams{{slot: 38, hasBlobs: true}},
-			storedDataColumns: []map[uint64]bool{{38: true, 102: true}},
-			peersParams: []peerParams{
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 38,
-							Count:     1,
-							Columns:   []uint64{6, 70},
-						}).String(): {
-							{},
-							{
-								{slot: 38, columnIndex: 6},
-								{slot: 38, columnIndex: 70},
-							},
-						},
-					},
-				},
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{{6, 70}},
-			isError:            false,
-		},
-		{
-			name:              "Some blocks with blobs with missing data columns - no response at all",
-			fuluForkEpoch:     1,
-			currentSlot:       40,
-			blocksParams:      []blockParams{{slot: 38, hasBlobs: true}},
-			storedDataColumns: []map[uint64]bool{{38: true, 102: true}},
-			peersParams: []peerParams{
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 38,
-							Count:     1,
-							Columns:   []uint64{6, 70},
-						}).String(): {{}, {}, {}, {}, {}, {}, {}, {}, {}, {}},
-					},
-				},
-			},
-			batchSize:          32,
-			addedRODataColumns: [][]int{{}},
-			isError:            true,
-		},
-		{
-			name:          "Some blocks with blobs with missing data columns - request has to be split",
-			fuluForkEpoch: 1,
-			currentSlot:   40,
-			blocksParams: []blockParams{
-				{slot: 32, hasBlobs: true}, {slot: 33, hasBlobs: true}, {slot: 34, hasBlobs: true}, {slot: 35, hasBlobs: true}, // 4
-				{slot: 36, hasBlobs: true}, {slot: 37, hasBlobs: true}, // 6
-			},
-			storedDataColumns: []map[uint64]bool{
-				nil, nil, nil, nil, // 4
-				nil, nil, // 6
-
-			},
-			peersParams: []peerParams{
-				{
-					cgc: 128,
-					toRespond: map[string][][]responseParams{
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 32,
-							Count:     4,
-							Columns:   []uint64{6, 38, 70, 102},
-						}).String(): {
-							{
-								{slot: 32, columnIndex: 6}, {slot: 32, columnIndex: 38}, {slot: 32, columnIndex: 70}, {slot: 32, columnIndex: 102},
-								{slot: 33, columnIndex: 6}, {slot: 33, columnIndex: 38}, {slot: 33, columnIndex: 70}, {slot: 33, columnIndex: 102},
-								{slot: 34, columnIndex: 6}, {slot: 34, columnIndex: 38}, {slot: 34, columnIndex: 70}, {slot: 34, columnIndex: 102},
-								{slot: 35, columnIndex: 6}, {slot: 35, columnIndex: 38}, {slot: 35, columnIndex: 70}, {slot: 35, columnIndex: 102},
-							},
-						},
-						(&ethpb.DataColumnSidecarsByRangeRequest{
-							StartSlot: 36,
-							Count:     2,
-							Columns:   []uint64{6, 38, 70, 102},
-						}).String(): {
-							{
-								{slot: 36, columnIndex: 6}, {slot: 36, columnIndex: 38}, {slot: 36, columnIndex: 70}, {slot: 36, columnIndex: 102},
-								{slot: 37, columnIndex: 6}, {slot: 37, columnIndex: 38}, {slot: 37, columnIndex: 70}, {slot: 37, columnIndex: 102},
-							},
-						},
-					},
-				},
-			},
-			batchSize: 4,
-			addedRODataColumns: [][]int{
-				{6, 38, 70, 102}, // Slot 32
-				{6, 38, 70, 102}, // Slot 33
-				{6, 38, 70, 102}, // Slot 34
-				{6, 38, 70, 102}, // Slot 35
-				{6, 38, 70, 102}, // Slot 36
-				{6, 38, 70, 102}, // Slot 37
-			},
-			isError: false,
-		},
-	}
-
-	// Initialize the trusted setup.
-	err := kzg.Start()
-	require.NoError(t, err)
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Consistency checks.
-			require.Equal(t, len(tc.blocksParams), len(tc.addedRODataColumns))
-
-			// Create a context.
-			ctx := context.Background()
-
-			// Create blocks, RO data columns and data columns sidecar from slot.
-			roBlocks := make([]blocks.ROBlock, len(tc.blocksParams))
-			roDatasColumns := make([][]blocks.RODataColumn, len(tc.blocksParams))
-			dataColumnsSidecarBySlot := make(map[primitives.Slot][]*ethpb.DataColumnSidecar, len(tc.blocksParams))
-
-			for i, blockParams := range tc.blocksParams {
-				pbSignedBeaconBlock := util.NewBeaconBlockFulu()
-				pbSignedBeaconBlock.Block.Slot = blockParams.slot
-
-				if blockParams.hasBlobs {
-					blobs := make([]kzg.Blob, blobsCount)
-					blobKzgCommitments := make([][]byte, blobsCount)
-
-					for j := range blobsCount {
-						blob := getRandBlob(t, int64(i+j))
-						blobs[j] = blob
-
-						blobKzgCommitment, err := kzg.BlobToKZGCommitment(&blob)
-						require.NoError(t, err)
-
-						blobKzgCommitments[j] = blobKzgCommitment[:]
-					}
-
-					pbSignedBeaconBlock.Block.Body.BlobKzgCommitments = blobKzgCommitments
-					signedBeaconBlock, err := blocks.NewSignedBeaconBlock(pbSignedBeaconBlock)
-					require.NoError(t, err)
-
-					cellsAndProofs := util.GenerateCellsAndProofs(t, blobs)
-					pbDataColumnsSidecar, err := peerdas.DataColumnSidecars(signedBeaconBlock, cellsAndProofs)
-					require.NoError(t, err)
-
-					dataColumnsSidecarBySlot[blockParams.slot] = pbDataColumnsSidecar
-
-					roDataColumns := make([]blocks.RODataColumn, 0, len(pbDataColumnsSidecar))
-					for _, pbDataColumnSidecar := range pbDataColumnsSidecar {
-						roDataColumn, err := blocks.NewRODataColumn(pbDataColumnSidecar)
-						require.NoError(t, err)
-
-						roDataColumns = append(roDataColumns, roDataColumn)
-					}
-
-					roDatasColumns[i] = roDataColumns
-				}
-
-				roDatasColumns = append(roDatasColumns, nil)
-				signedBeaconBlock, err := blocks.NewSignedBeaconBlock(pbSignedBeaconBlock)
-				require.NoError(t, err)
-
-				roBlock, err := blocks.NewROBlock(signedBeaconBlock)
-				require.NoError(t, err)
-
-				roBlocks[i] = roBlock
-			}
-
-			// Set the Fulu fork epoch.
-			params.BeaconConfig().FuluForkEpoch = tc.fuluForkEpoch
-
-			// Save the blocks in the store.
-			storage := make(map[[fieldparams.RootLength]byte][]uint64)
-			for index, columns := range tc.storedDataColumns {
-				root := roBlocks[index].Root()
-
-				columnsSlice := make([]uint64, 0, len(columns))
-				for column := range columns {
-					columnsSlice = append(columnsSlice, column)
-				}
-
-				storage[root] = columnsSlice
-			}
-
-			dataColumnStorageSummarizer := filesystem.NewMockDataColumnStorageSummarizer(t, storage)
-
-			// Create a chain and a clock.
-			chain, clock := defaultMockChain(t, tc.currentSlot)
-
-			// Create the P2P service.
-			p2pSvc := p2ptest.NewTestP2P(t, libp2p.Identity(genFixedCustodyPeer(t)))
-			nodeID, err := p2p.ConvertPeerIDToNodeID(p2pSvc.PeerID())
-			require.NoError(t, err)
-			p2pSvc.EnodeID = nodeID
-
-			// Connect the peers.
-			peers := make([]*p2ptest.TestP2P, 0, len(tc.peersParams))
-			for i, peerParams := range tc.peersParams {
-				peer := createAndConnectPeer(t, p2pSvc, chain, dataColumnsSidecarBySlot, peerParams, i)
-				peers = append(peers, peer)
-			}
-
-			peersID := make([]peer.ID, 0, len(peers))
-			for _, peer := range peers {
-				peerID := peer.PeerID()
-				peersID = append(peersID, peerID)
-			}
-
-			status := &ethpb.Status{HeadSlot: peersHeadSlot}
-
-			for _, peerID := range peersID {
-				p2pSvc.Peers().SetChainState(peerID, status)
-			}
-
-			clockSync := startup.NewClockSynchronizer()
-			require.NoError(t, clockSync.SetClock(clock))
-			iniWaiter := verification.NewInitializerWaiter(clockSync, nil, nil)
-			ini, err := iniWaiter.WaitForInitializer(ctx)
-			require.NoError(t, err)
-
-			// Create the block fetcher.
-			blocksFetcher := newBlocksFetcher(ctx, &blocksFetcherConfig{
-				clock:       clock,
-				ctxMap:      map[[4]byte]int{{245, 165, 253, 66}: version.Fulu},
-				p2p:         p2pSvc,
-				dcs:         dataColumnStorageSummarizer,
-				cv:          newDataColumnsVerifierFromInitializer(ini),
-				custodyInfo: &peerdas.CustodyInfo{},
-			})
-
-			// Fetch the data columns from the peers.
-			fetchedRoDataColumnsByRoot, err := blocksFetcher.fetchMissingDataColumnsFromPeers(ctx, roBlocks, peersID, tc.batchSize)
-			if !tc.isError {
-				require.NoError(t, err)
-			} else {
-				require.NotNil(t, err)
-				return
-			}
-
-			expectedDataColumnsByRoot := make(map[[fieldparams.RootLength]byte][]blocks.RODataColumn)
-
-			for i, addedColumns := range tc.addedRODataColumns {
-				root := roBlocks[i].Root()
-				expectedRODataColumns := make([]blocks.RODataColumn, 0, len(tc.addedRODataColumns[i]))
-				for _, column := range addedColumns {
-					roDataColumn := roDatasColumns[i][column]
-					expectedRODataColumns = append(expectedRODataColumns, roDataColumn)
-				}
-
-				if len(expectedRODataColumns) > 0 {
-					expectedDataColumnsByRoot[root] = expectedRODataColumns
-				}
-			}
-
-			require.Equal(t, len(expectedDataColumnsByRoot), len(fetchedRoDataColumnsByRoot))
-
-			for root := range expectedDataColumnsByRoot {
-				expectedDataColumns := expectedDataColumnsByRoot[root]
-				fetchedDataColumns := fetchedRoDataColumnsByRoot[root]
-
-				sort.Slice(expectedDataColumns, func(i, j int) bool {
-					return expectedDataColumns[i].Index < expectedDataColumns[j].Index
-				})
-
-				sort.Slice(fetchedDataColumns, func(i, j int) bool {
-					return fetchedDataColumns[i].Index < fetchedDataColumns[j].Index
-				})
-
-				require.DeepSSZEqual(t, expectedDataColumns, fetchedDataColumns)
-			}
-		})
-	}
-}
-
-// This generates a peer which custodies the columns of 6,38,70 and 102.
-func genFixedCustodyPeer(t *testing.T) crypto.PrivKey {
-	rawObj, err := hex.DecodeString("58f40e5010e67d07e5fb37c62d6027964de2bef532acf06cf4f1766f5273ae95")
-	if err != nil {
-		t.Fatal(err)
-	}
-	pkey, err := crypto.UnmarshalSecp256k1PrivateKey(rawObj)
-	require.NoError(t, err)
-	return pkey
 }
