+++ beacon-chain/sync/initial-sync/blocks_fetcher_utils.go
@@ -3,7 +3,6 @@ package initialsync
 import (
 	"context"
 	"fmt"
-	"slices"
 
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/pkg/errors"
@@ -296,7 +295,6 @@ func (f *blocksFetcher) findForkWithPeer(ctx context.Context, pid peer.ID, peers
 
 // findAncestor tries to figure out common ancestor slot that connects a given root to known block.
 func (f *blocksFetcher) findAncestor(ctx context.Context, pid peer.ID, peers []peer.ID, b interfaces.ReadOnlySignedBeaconBlock) (*forkData, error) {
-
 	outBlocks := []interfaces.ReadOnlySignedBeaconBlock{b}
 	for i := uint64(0); i < backtrackingMaxHops; i++ {
 		parentRoot := outBlocks[len(outBlocks)-1].Block().ParentRoot()
@@ -358,79 +356,4 @@ func (f *blocksFetcher) calculateHeadAndTargetEpochs() (headEpoch, targetEpoch p
 	targetEpoch, peers = f.p2p.Peers().BestNonFinalized(flags.Get().MinimumSyncPeers, headEpoch)
 
 	return headEpoch, targetEpoch, peers
-}
-
-// uint64MapToSortedSlice produces a sorted uint64 slice from a map.
-func uint64MapToSortedSlice(input map[uint64]bool) []uint64 {
-	output := make([]uint64, 0, len(input))
-	for idx := range input {
-		output = append(output, idx)
-	}
-
-	slices.Sort[[]uint64](output)
-	return output
-}
-
-// Filter peers with head epoch lower than our target epoch for ByRange requests.
-func (f *blocksFetcher) filterPeersByTargetSlot(peers []peer.ID, targetSlot primitives.Slot) ([]peer.ID, []string, error) {
-	filteredPeers := make([]peer.ID, 0, len(peers))
-	descriptions := make([]string, 0, len(peers))
-	// Compute the target epoch from the target slot.
-	targetEpoch := slots.ToEpoch(targetSlot)
-
-	for _, peer := range peers {
-		peerChainState, err := f.p2p.Peers().ChainState(peer)
-		if err != nil {
-			description := fmt.Sprintf("peer %s: error: %s", peer, err)
-			descriptions = append(descriptions, description)
-			continue
-		}
-
-		if peerChainState == nil {
-			description := fmt.Sprintf("peer %s: chain state is nil", peer)
-			descriptions = append(descriptions, description)
-			continue
-		}
-
-		peerHeadEpoch := slots.ToEpoch(peerChainState.HeadSlot)
-
-		if peerHeadEpoch < targetEpoch {
-			description := fmt.Sprintf("peer %s: peer head epoch %d < our target epoch %d", peer, peerHeadEpoch, targetEpoch)
-			descriptions = append(descriptions, description)
-			continue
-		}
-
-		filteredPeers = append(filteredPeers, peer)
-	}
-
-	return filteredPeers, descriptions, nil
-}
-
-// Filter peers to ensure they are synced to the target slot and have sufficient bandwidth to serve the request.
-func (f *blocksFetcher) filterPeersByTargetSlotAndBandwidth(peers []peer.ID, lastSlot primitives.Slot, blockCount uint64) ([]peer.ID, []string, error) {
-	if len(peers) == 0 {
-		peers = f.p2p.Peers().Connected()
-	}
-
-	slotPeers, descriptions, err := f.filterPeersByTargetSlot(peers, lastSlot)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "peers with slot and data columns")
-	}
-
-	// Filter for peers with sufficient bandwidth to serve the request.
-	slotAndBandwidthPeers := f.hasSufficientBandwidth(slotPeers, blockCount)
-
-	// Add debugging logs for the filtered peers.
-	peerWithSufficientBandwidthMap := make(map[peer.ID]bool, len(peers))
-	for _, peer := range slotAndBandwidthPeers {
-		peerWithSufficientBandwidthMap[peer] = true
-	}
-
-	for _, peer := range slotPeers {
-		if !peerWithSufficientBandwidthMap[peer] {
-			description := fmt.Sprintf("peer %s: does not have sufficient bandwidth", peer)
-			descriptions = append(descriptions, description)
-		}
-	}
-	return slotAndBandwidthPeers, descriptions, nil
 }
