+++ beacon-chain/sync/rpc_blob_sidecars_by_root_test.go
@@ -190,7 +190,7 @@ func TestBlobsByRootValidation(t *testing.T) {
 	}()
 	capellaSlot, err := slots.EpochStart(params.BeaconConfig().CapellaForkEpoch)
 	require.NoError(t, err)
-	dmc, clock := defaultMockChain(t)
+	dmc, clock := defaultMockChain(t, 0)
 	dmc.Slot = &capellaSlot
 	dmc.FinalizedCheckPoint = &ethpb.Checkpoint{Epoch: params.BeaconConfig().CapellaForkEpoch}
 	cases := []*blobsTestCase{
